Move helpers to utils file
Enact-DCO-1.0-Signed-off-by: Roy Sutton roy.sutton@lge.com
