+++ test/moonstone/ExpandableList/ExpandableList-specs.js
@@ -1,4 +1,5 @@
-let Page = require('./ExpandableListPage');
+const Page = require('./ExpandableListPage'),
+	{validateTitle, expectClosed, expectOpen} = require('./ExpandableList-utils.js');
 
 describe('ExpandableList', function () {
 
@@ -537,23 +538,3 @@ describe('ExpandableList', function () {
 		});
 	});
 });
-
-// Validations are self-contained 'it' statements
-function validateTitle (expandable, title) {
-	it('should have correct title', function () {
-		expect(expandable.titleText).to.equal(title);
-	});
-}
-
-// Expects are blocks of expects or other commands to be embedded in an 'it' statement
-function expectClosed (expandable) {
-	expect(expandable.isOpen).to.be.false();
-	expect(expandable.chevron).to.equal('󯿭');
-	expect(expandable.item(0).isVisible()).to.be.false();
-}
-
-function expectOpen (expandable) {
-	expect(expandable.isOpen).to.be.true();
-	expect(expandable.chevron).to.equal('󯿮');
-	expect(expandable.item(0).isVisible()).to.be.true();
-}
